Makes the use of integer types more consistent.
Uses the integer types like i16 or u32 when an API takes integers with defined size (e.g. int32_t or i32).
For unsigned int it uses uint. I've moved that typedef to a more suitable location.
Also changes the conversion for long and unsigned long because SpiderMonkey now has conversion functions for 64 bit integers.
It's still not quite consistent from the perspective of just looking at a file and seeing that some loops use int and some u32 as counter, but casts for the function calls wouldn't be better.

+++ source/graphics/MapGenerator.cpp
@@ -106,14 +106,14 @@ bool CMapGeneratorWorker::Run()
 		else
 		{
 			// Init RNG seed
-			uint32_t seed;
+			u32 seed;
 			if (!m_ScriptInterface->GetProperty(settingsVal.get(), "Seed", seed))
 			{	// No seed specified
 				LOGWARNING(L"CMapGeneratorWorker::Run: No seed value specified - using 0");
 				seed = 0;
 			}
 
-			m_MapGenRNG.seed((int32_t)seed);
+			m_MapGenRNG.seed(seed);
 
 			// Copy settings to global variable
 			if (!m_ScriptInterface->SetProperty(m_ScriptInterface->GetGlobalObject(), "g_MapSettings", settingsVal))
